typescript : defaults to arraybuffer when content-type is unknown #24
