+++ lib/contextBuilder.js
@@ -1,4 +1,3 @@
-/// <reference path="../typings.d.ts" />
 var gutil = require('gulp-util');
 var _ = require('lodash');
 var typeMapperInstance;
@@ -90,6 +89,7 @@ var Operation = (function () {
         }
         this.isJsonRequest = this.consumes.filter(function (x) { return x === "application/json"; }).length > 0;
         this.isJsonResponse = this.produces.filter(function (x) { return x === "application/json"; }).length > 0;
+        this.isBinaryResponse = !this.isJsonResponse;
         this.isFormDataRequest = this.consumes.filter(function (x) { return x === "multipart/form-data"; }).length > 0;
         this.security = method.security ? _.keys(method.security[0])[0] : null;
         _.forEach(method.responses, function (response, status) {
