+++ lib/diagnostics.js
@@ -1,4 +1,3 @@
-/// <reference path="../typings.d.ts" />
 var gutil = require('gulp-util');
 exports.PLUGIN_NAME = 'gulp-swagger-generator';
 function newPluginError(message) {
