+++ lib/index.js
@@ -1,4 +1,3 @@
-/// <reference path="../typings.d.ts" />
 var through = require('through2');
 var diagnostics = require('./diagnostics');
 var parse = require('./parse');
