+++ lib/parse.js
@@ -1,4 +1,3 @@
-/// <reference path="../typings.d.ts" />
 var Promise = require('bluebird');
 var parser = require("swagger-parser");
 var gutil = require("gulp-util");
