+++ lib/typescriptTypeMapper.js
@@ -1,4 +1,3 @@
-/// <reference path="../typings.d.ts" />
 var gutil = require('gulp-util');
 var _ = require('lodash');
 function createTypeMapper(context) {
