+++ package.json
@@ -1,6 +1,6 @@
 {
  "name": "gulp-swagger-generator",
- "version": "0.0.196",
+ "version": "0.0.204",
  "description": "",
  "license": "",
  "repository": "",
@@ -26,12 +26,13 @@
  },
  "devDependencies": {
   "gulp": "^3.9.0",
-  "gulp-bump": "^0.3.1",
-  "gulp-help": "^1.6.0",
-  "gulp-jasmine": "^2.0.1",
-  "gulp-tslint": "^3.0.2-beta",
-  "jasmine": "^2.3.1",
-  "tslint": "2.*"
+  "gulp-bump": "^1.0.0",
+  "gulp-help": "^1.6.1",
+  "gulp-jasmine": "^2.2.1",
+  "gulp-tslint": "^4.3.1",
+  "jasmine": "^2.4.1",
+  "tslint": "^3.3.0",
+  "typescript": "^1.7.5"
  },
  "engines": {
   "node": ">=0.10.0"
