+++ src/contextBuilder.ts
@@ -112,6 +112,7 @@ export class Operation {
 
     public isJsonRequest: boolean;
     public isJsonResponse: boolean;
+    public isBinaryResponse: boolean;
     public isFormDataRequest: boolean;
 
     public description: string;
@@ -170,6 +171,7 @@ export class Operation {
 
         this.isJsonRequest = this.consumes.filter(x=> x === "application/json").length > 0;
         this.isJsonResponse = this.produces.filter(x=> x === "application/json").length > 0;
+        this.isBinaryResponse = !this.isJsonResponse;
         this.isFormDataRequest = this.consumes.filter(x=> x === "multipart/form-data").length > 0;
 
         this.security = method.security ? _.keys(method.security[0])[0] : null;
